requirements.txt
numpy
scikit-learn
matplotlib
python-chess
stockfish
tqdm
flask